Moved g2a forms and views to partners
